feat: ocultar listas de usuarios y empleados para roles no administrativos
- Se añadió la propiedad `isHidden` en las listas de `User` y `Employee`, asegurando que solo los usuarios con rol 'ADMIN' puedan visualizarlas.
- Se actualizó la sección de navegación personalizada para eliminar el rol 'NORMAL' de la opción 'Captura Semanal', permitiendo acceso solo a roles 'CAPTURA' y 'ADMIN'.
- Se implementó un nuevo resolver `getFinancialReport` para optimizar la generación de reportes financieros, mejorando la gestión de datos y la eficiencia de las consultas.
